Update world_generator.py
Implements the ability the have a list of desirable materials for the agent to find. A future update will be to have a dictionary/mapping detailing how many are required of each, thus allowing the agent to keep track of only what is necessary, rather than searching for all materials all the time.